rename time_module.rs to time.rs
This commit resolve issue #2882.
* rename time_module.rs to time.rs
